refactor: 화면 입출력 요소를 InputView, OutputView의 두 클래스로 관리하도록 변경
