+++ src/main/java/subway/views/InputView.java
@@ -1,19 +1,21 @@
-package subway.views.mainviews;
+package subway.views;
+
+import subway.menus.MainMenu;
 
 import java.util.Arrays;
 import java.util.Scanner;
 
-public class MainInputView {
+public class InputView {
     private static final String NOT_EXIST_MAIN_MENU_SELECTION = "[ERROR] 선택할 수 없는 기능입니다.";
+
+    private static String userInput(Scanner scanner) {
+        return scanner.nextLine();
+    }
 
     public static String selectMainMenu(Scanner scanner) {
         String selection = userInput(scanner);
         checkExistMainMenu(selection);
         return selection;
-    }
-
-    private static String userInput(Scanner scanner) {
-        return scanner.nextLine();
     }
 
     private static void checkExistMainMenu(String selection) {
@@ -24,7 +26,7 @@ public class MainInputView {
 
     private static boolean isExistMainMenu(String selection) {
         boolean isExist = Arrays.stream(MainMenu.values())
-                            .anyMatch(menu -> menu.getOptionalItem().equals(selection));
+            .anyMatch(menu -> menu.getOptionalItem().equals(selection));
         return isExist;
     }
 }
